Reconfigured layout
Since Transaction History is like a runtime stack that grows downward (while heap grows upward), I allocated it towards the bottom where it can grow to fit the page.

The form that adds the transaction should be visible and remain static for the user so it was placed at the top.

Also center the balance so it serves as the focal point for the user.